financial-analyzer/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.7.0
openai-whisper==20231117
opencv-python==4.8.1.78
pytesseract==0.3.10
python-dotenv==1.0.0
